Add extra fields to struct serverThreadState
Add the numHosts and serverNetLoad extra fields to the struct
serverThreadState. The first is the same as in struct
clientThreadState and the latter determines the numbers of messages
to be sent to other servers when serving a HP client request.

Also, create a continuation struct which will be used to determine
whether all the required server-to-server communication is completed
before sending a response to a client. Each server thread has an array
of such structs. A bitVector is used to find a free entry in the
struct and allocate it for a specific client request.

Signed-off-by: Kostis Kaffes <kkaffes@stanford.edu>