Add a helper functions to the server
Add the serverChooseServerID function so that a server can identify to
which other servers it should send messages while it serves a client
request.

Add the serverChooseServerRank function to choose the rank of the
server to send the S2S request to.

Add the serverCreateNetworkRequest function to send network requests
to other servers according to the serverNetLoad parameter. Select a
different (random) target server for each request.

All the above functions are almost duplicate of the client functions.
If possible in the future, merge them.

Signed-off-by: Kostis Kaffes <kkaffes@stanford.edu>