Added a hostfile.ompi which tells mpirun which hosts to launch MPI processes onto.
Simplified run_sim.py because we are not going to be running MVAPICH.
Also changed the mpirun options to run on Ethernet instead of
infiniband.

The bug with the 'ERROR: current_time - ipg < 0' was because the
clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &time) gives the time since
the process started running on the CPU. So if either the HPMsgRate
or the LPMsgRate is very low then ipg will be very high and may be
higher than the current_time.

Created a script 'transfer_to_maverick.sh' which copies the project
over to maverick-1 and rebuilds it over there.