Track s2s latency
Track and plot the latency of S2S HP requests. Use the continuation
that tracks the replies and add a field that holds the time the first
request is sent out. After each reply arrives, the server notes the
current time and their difference is the latency of server2server
communication.

Make minor changes in hdr_histogram so that it is easier to do pattern
matching.

Signed-off-by: Kostis Kaffes <kkaffes@stanford.edu>